Adapt to changes in Qt 5.9 git
A recent commit upstream [1] renamed some configure tests which resulted
in a broken build. It seems that these sed invocations were effectively
a no-op since at least [2] which apparently removed the last instance of
qmake's `-nocache` argument.

This seding was in place in the qt5-build.eclass since its very
beginning. I believe that the explanatory comment does not effectively
match what the code (might have been trying to) do. It seems that the
real origin dates back to 2015 when it was added [3] to the
qt4-build-multilib.eclass by @pesa.

Given that this `sed` was not doing anything for the past year anyway
and that nobody reported a bug about that AFAIK and that it started
breaking the build, it is safe to remove it now.

[1] https://code.qt.io/cgit/qt/qtbase.git/commit/?id=b0060d1056d6d1752d91652261de97db909c7862
[2] https://code.qt.io/cgit/qt/qtbase.git/commit/?id=d90db0f136d727f8e6b9a566eee3eae80cd7d81e
[3] https://github.com/gentoo/gentoo-gitmig-20150809-draft/commit/265a3c916261b238bc12505a031d92225b1d7056